making a mast base class
